Renamed userReducer to reducer, since we're only going to have one
